+++ Engine/Controls/AiServiceItemControl.xaml.cs
@@ -161,6 +161,7 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 			if (ControlsHelper.AllowLoad(this))
 			{
 				AppHelper.InitHelp(this);
+				UiPresetsManager.InitControl(this);
 			}
 		}