+++ Engine/Controls/AssistantItemControl.xaml.cs
@@ -147,6 +147,7 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 			if (ControlsHelper.AllowLoad(this))
 			{
 				AppHelper.InitHelp(this);
+				UiPresetsManager.InitControl(this);
 			}
 		}
 	}