+++ Engine/Controls/Options/MicrosoftAccountsControl.xaml.cs
@@ -309,6 +309,7 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls.Options
 				var profile = MicrosoftAccountManager.Current.GetProfile();
 				profile.PropertyChanged += Profile_PropertyChanged;
 				AppHelper.InitHelp(this);
+				UiPresetsManager.InitControl(this, true);
 			}
 		}