+++ Engine/Controls/PropertyComboBox.xaml.cs
@@ -19,6 +19,7 @@ namespace JocysCom.VS.AiCompanion.Engine
 			if (ControlsHelper.AllowLoad(this))
 			{
 				AppHelper.InitHelp(this);
+				UiPresetsManager.InitControl(this, true);
 			}
 		}
 	}