+++ Engine/Controls/TemplateItemControl.xaml.cs
@@ -513,6 +513,9 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 					);
 				}
 				AppHelper.InitHelp(this);
+				UiPresetsManager.InitControl(this);
+				// Remove control, which visibility is controlled by the code.
+				UiPresetsManager.RemoveControls(this);
 			}
 			RestoreFocus();
 			UpdateAvatarControl();