+++ Engine/Controls/VaultItemControl.xaml.cs
@@ -115,6 +115,7 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 				var profile = MicrosoftAccountManager.Current.GetProfile();
 				profile.PropertyChanged += Profile_PropertyChanged;
 				AppHelper.InitHelp(this);
+				UiPresetsManager.InitControl(this);
 			}
 		}