+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix: List not expanding.
 - Fix: Improved support for Azure SQL Server Databases.
 - Fix: Unable to stop message requests.
+- Fix: Embedding search group filter not working correctly.
 - Update: SQL now uses native SQL script AI search instead of C# Assembly.
 - Update: UI, Templates and AI Avatar updated.
 - New: Create UI presets, like "Simple" or "Advanced".